+++ template/src/services/cron.service/index.ts
@@ -1,0 +1,18 @@
+import { demoCronJob } from './demo.job';
+import _ from 'lodash';
+import { BaseCronJob } from './base.job';
+
+export const cronJobs = [
+    demoCronJob
+];
+
+/**
+ * 启动定时任务
+ */
+export const startCronJobs = () => {
+    _.forEach(cronJobs, (cronJob: BaseCronJob) => {
+        if (!cronJob.isAutoStart) {
+            cronJob.job.start();
+        }
+    });
+};